requirements.txt
alembic==1.8.1
anyio==3.6.1
asyncpg==0.26.0
autopep8==1.6.0
click==8.1.3
cffi==1.15.1
autopep8==1.6.0
cryptography==37.0.4
databases==0.6.0
ecdsa==0.18.0
fastapi==0.79.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
Mako==1.2.1
pyasn1==0.4.8
MarkupSafe==2.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
rsa==4.8
pydantic==1.9.1
python-dotenv==0.20.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3
